+++ docs/modules/TreeFormatter.ts.md
@@ -1,6 +1,6 @@
 ---
 title: TreeFormatter.ts
-nav_order: 8
+nav_order: 7
 parent: Modules
 ---
